requirements.txt
numpy==1.14.5
opencv-python==4.4.0
tqdm==4.51.0
scikit-image==0.17.1
Pillow==8.0.1
scipy==1.2.1
